delete and get actions for make/take survey
Add delete-survey event in make survey and get-suveys/survey event in take survey events file

+++ assets/scripts/makesurvey/api.js
@@ -24,14 +24,17 @@ const updateSurvey = (data, survey_id) => {
   });
 };
 
-const deleteSurvey = (survey_id) => {
-  return $.ajax({
-    url: app.host + '/surveys/' + survey_id,
-    method: 'DELETE',
-    headers: {
-      Authorization: 'Token token=' + app.user.token,
-      },
-});
+//delete function need to test, because we don't know what is app.user.survey.id is.
+const deleteSurvey = (data) => {
+  if (data === app.user.survey.title){
+    return $.ajax({
+      url: app.host + '/surveys/' + app.user.survey.id,
+      method: 'DELETE',
+      headers: {
+        Authorization: 'Token token=' + app.user.token,
+        },
+   });
+  }
 };
